Fixed sign comparison errors caught by hardening flags in stout.
The hardening flags produced many new sign comparison errors
in stout that need to be fixed for Mesos to compile/run.

Review: https://reviews.apache.org/r/52886/

+++ 3rdparty/stout/tests/ip_tests.cpp
@@ -47,8 +47,8 @@ TEST_TEMP_DISABLED_ON_WINDOWS(NetTest, LinkDevice)
     if (network.isSome()) {
       string addr = stringify(network.get());
       string prefix = addr.substr(addr.find("/") + 1);
-      ASSERT_SOME(numify<size_t>(prefix));
-      EXPECT_EQ(network.get().prefix(), numify<size_t>(prefix).get());
+      ASSERT_SOME(numify<int>(prefix));
+      EXPECT_EQ(network.get().prefix(), numify<int>(prefix).get());
 
       vector<string> tokens =
         strings::split(addr.substr(0, addr.find("/")), ".");
